Added routes folder and split the routes into two files with a helper function file to make code cleaner and easier to read. moved original routes.py file into routes folder and commented out original code for reference of beginning code.
