+++ app/__init__.py
@@ -31,8 +31,8 @@ def create_app(test_config=None):
     migrate.init_app(app, db)
 
     # Register Blueprints here
-    from .routes import task_bp
+    from .routes.task import task_bp
     app.register_blueprint(task_bp)
-    from .routes import goals_bp
+    from .routes.goal import goals_bp
     app.register_blueprint(goals_bp)
     return app
